+++ CHANGELOG.md
@@ -27,6 +27,6 @@ hi
 - Cracked and Chiseled Purpur Blocks can now be broken faster with a pickaxe. ([#10](https://github.com/FrozenBlock/TrailierTales/issues/10))
 - Fixed an issue where modded boats that aren't implemented the same way as vanilla cause a crash. ([#9](https://github.com/FrozenBlock/TrailierTales/issues/9))
 - Fixed the Polished Calcite Wall not being craftable with a Stonecutter. ([#11](https://github.com/FrozenBlock/TrailierTales/issues/11))
-- Fixed many recipe unlocks not working on 1.21.2+.
+- Fixed some recipe unlocks not working.
 - Removed the leftover `TRAILIER TALES` subtitle from the main menu.
 - Removed many unused assets.